agentcore/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bedrock-agentcore>=0.1.1",
    "bedrock-agentcore-starter-toolkit==0.1.2",
    "strands-agents>=1.2.0",
    "strands-agents-tools>=0.2.3",
    "streamlit>=1.47.1",
]
